% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates-miss.R
\docType{data}
\name{dates_miss}
\alias{dates_miss}
\title{Simulated data with dates missing}
\format{
\code{dates_miss} is supplied with the \pkg{lhmetools} package and
consists of the following information in a data frame with 10 rows and 4 columns.
\describe{
\item{id}{station id}
\item{g}{group, e.g country or region id}
\item{date}{date of observation}
\item{value}{observed value of some meteorological variable}
}
}
\usage{
dates_miss
}
\description{
This data mimics a typical data set of a weather station.
}
\examples{
str(dates_miss)
}
\keyword{datasets}
