% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhmetools-package.R
\docType{package}
\name{lhmetools}
\alias{lhmetools-package}
\alias{lhmetools}
\title{lhmetools: convenience functions used in LHMET}
\description{
The goal of \pkg{lhmetools} is to bundle all generic functions frequently
used in LHMET.
}
\section{lhmetools functions}{

So far now the included functions are:
\itemize{
\item \code{\link{unrar}}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lhmet/lhmetools}
}

}
\author{
\strong{Maintainer}: Jônatan Tatsch \email{lhmet@ufsm.br}

}
