% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-clean.R
\name{str_sanitize}
\alias{str_sanitize}
\title{Sanitize strings by removing reserved and non portable character
set}
\usage{
str_sanitize(vnames, sep = "_", no.accent = TRUE)
}
\arguments{
\item{vnames}{character vector to be sanitized}

\item{sep}{character of replacements, Default: '_'}

\item{no.accent}{logical, if is to remove accents from \code{vnames},
Default: TRUE}
}
\value{
a character vector in lowercase with underscore (sep) to separate
nouns
}
\description{
Make a reasonable attempt of converting a string into a
preferred standard form to name variables and file names
}
\details{
\code{str_sanitize()} removes or replace the following:
\itemize{
\item \href{https://en.wikipedia.org/wiki/C0_and_C1_control_codes}{Control characters}
\item \href{https://kb.acronis.com/content/39790}{Reserved characters}
\item Unix reserved filenames (\code{.} and \code{..})
\item Trailing periods and spaces (invalid on Windows)
\item Windows reserved filenames (\code{CON}, \code{PRN}, \code{AUX}, \code{NUL}, \code{COM1}, \code{COM2},
\code{COM3}, COM4, \code{COM5}, \code{COM6}, \code{COM7}, \code{COM8}, \code{COM9}, \code{LPT1}, \code{LPT2},
\code{LPT3}, \code{LPT4}, \code{LPT5}, \code{LPT6}, LPT7, \code{LPT8}, and \code{LPT9})
\item any accented noum and punctuation character
\item any resulting initial or trailing underscore or multiples
\item uppercase by lowercase
\item repeated seperator
\item names starting with a number by a character
}
}
\examples{
str_sanitize(c("esúpido", "^ ãb ", "..c`a§", "A .xls.xls", "1° dia"))
}
