% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unrar.R
\name{unrar}
\alias{unrar}
\title{Extract files from \code{rar} archives (only works on Linux)}
\usage{
unrar(
  file,
  dest_dir = fs::path_ext_remove(file),
  overwrite = FALSE,
  quiet = TRUE
)
}
\arguments{
\item{file}{a file path to a \code{file.rar}}

\item{dest_dir}{path to extract files. Defaults to \code{dirname(file)}.}

\item{overwrite}{logical, use \code{overwrite = TRUE} to overwrite
existing files.}

\item{quiet}{Hide printed output, messages, warnings, and errors
(TRUE, the default), or display them as they occur?}
}
\value{
character vector with files path
}
\description{
This function extract files from a \code{.rar} file
}
\details{
This function has the side effect of extract files in a
directory named (\code{dest_dir/basename(file)}) when \code{dest_dir} is not \code{NULL}.
}
\note{
This function is a wrapper to call \code{unrar} command. It is required you
have \code{unrar} Linux library installed. You can install it with
\verb{apt install unrar}.
}
\examples{
\dontrun{
if (interactive()) {

  rarfile_url <- paste0("https://www.dropbox.com/scl/fi/63eimy6j2ok1q3vf11zs4/",
  "some-file.rar?rlkey=jhveob9ysl1ivy3alnt68c61g&st=dvt43nb1&dl=1")

  dest_file <- file.path(tempdir(), strsplit(basename(rarfile_url), "\\\\?")[[1]][[1]])
  #dest_file <- tempfile(fileext = ".rar")
  download.file(rarfile_url, dest_file, mode = "wb")
  extracted_files <- unrar(dest_file)
  extracted_files
}
}
}
