% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates-manipulation.R
\name{dates_from_files}
\alias{dates_from_files}
\title{Extract dates from file names}
\usage{
dates_from_files(x, quietly = TRUE)
}
\arguments{
\item{x}{charcter}

\item{quietly}{logical. If TRUE, function evaluates without displaying
customary messages.}
}
\value{
a vector of class Date
}
\description{
Extract dates from file names
}
\examples{
dates_from_files("MERGE_CPTEC_20201201.grib2")
dates_from_files("brick-ETo-25km-19800101-20170731.nc")

}
\seealso{
Other date manipulation: 
\code{\link{complete_dates}()}
}
\concept{date manipulation}
